Rename textExtractFromObj module to textExtract
