+++ src/utils/textExtract.js
